mcp_servers/mcp-1c-hack/pyproject.toml
[project]
name = "find-path-mcp"
version = "0.1.0"
description = "Universal on-demand 1C MCP server with automatic dump, cache, and fuzzy search."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.2.0",
    "python-dotenv>=1.0.1",
    "opentelemetry-api>=1.25.0",
    "pydantic>=2.9.0",
    "platformdirs>=4.0.0",
    "thefuzz[speedup]>=0.22.1",
    "lxml>=5.3.0",
    "httpx>=0.27.0",
    "aiohttp>=3.10.0",
    "tenacity>=8.2.3",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
test = ["pytest>=8.3.0"]

[tool.setuptools]
py-modules = ["mcp_instance", "server", "manager", "parser", "config", "query_tool"]
packages = ["tools", "odata_tool"]
